+++ src/main/java/de/tilman_neumann/jml/factor/siqs/tdiv/TDiv_QS_3LP.java
@@ -253,8 +253,8 @@ public class TDiv_QS_3LP implements TDiv_QS {
 					xModP = x+p;
 				} else {
 					final long m = pinvArrayL[pIndex];
-					final long q = ((x*m)>>>32);
-					xModP = (int) (x - q * p);
+					final long q = ( ( ((long)x) * m) >>> 32); // first argument long optimizes register usage
+					xModP = (int) ( ((long)x) - q * p);
 					if (xModP<0) xModP += p;
 					if (DEBUG) {
 						// 0 <= xModP < p
@@ -290,8 +290,8 @@ public class TDiv_QS_3LP implements TDiv_QS {
 					xModP = x;
 				} else {
 					final long m = pinvArrayL[pIndex];
-					final long q = ((x*m)>>>32);
-					xModP = (int) (x - q * p);
+					final long q = ( ( ((long)x) * m) >>> 32); // first argument long optimizes register usage
+					xModP = (int) ( ((long)x) - q * p);
 					if (xModP>=p) xModP -= p;
 					if (DEBUG) {
 						// 0 <= xModP < p
